hf-spaces/alzheimers-model/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
gradio==3.41.2
python-multipart==0.0.6
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
requests==2.31.0
scikit-learn==1.3.0 